Converted all floats to doubles and converted missing values to zero

+++ GSOD-DataProcessor/Models/Non-EF Models/StationGSOD.cs
@@ -4,29 +4,29 @@ public class StationGSOD
 {
     public string? StationNumber { get; set; }
     public DateTime Date { get; set; }
-    public float Latitude { get; set; }
-    public float Longitude { get; set; }
-    public float Elevation { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
+    public double Elevation { get; set; }
     public string? StationName { get; set; }
-    public float Temp_Mean { get; set; }
+    public double Temp_Mean { get; set; }
     public string? Temp_Mean_Attribute { get; set; }
-    public float Dewp_Mean { get; set; }
+    public double Dewp_Mean { get; set; }
     public string? Dewp_Mean_Attribute { get; set; }
-    public float StationPress_Mean { get; set; }
+    public double StationPress_Mean { get; set; }
     public string? StationPress_Mean_Attribute { get; set; }
-    public float SeaLevelPress_Mean { get; set; }
+    public double SeaLevelPress_Mean { get; set; }
     public string? SeaLevelPress_Mean_Attribute { get; set; }
-    public float Visi_Mean { get; set; }
+    public double Visi_Mean { get; set; }
     public string? Visi_Mean_Attribute { get; set; }
-    public float WndSpd_Mean { get; set; }
+    public double WndSpd_Mean { get; set; }
     public string? WndSpd_Mean_Attribute { get; set; }
-    public float WndSpd_Max { get; set; }
-    public float WndSpd_Gust { get; set; }
-    public float Temp_Max { get; set; }
+    public double WndSpd_Max { get; set; }
+    public double WndSpd_Gust { get; set; }
+    public double Temp_Max { get; set; }
     public string? Temp_Max_Attribute { get; set; }
-    public float Temp_Min { get; set; }
+    public double Temp_Min { get; set; }
     public string? Temp_Min_Attribute { get; set; }
-    public float Precip_Total { get; set; }
+    public double Precip_Total { get; set; }
     public string? Precip_Total_Attribute { get; set; }
-    public float Snow_Depth { get; set; }
+    public double Snow_Depth { get; set; }
 }
